+++ modules/outbound-command-event-handler/test/integration/application/outbound_command_event_handler.test.ts
@@ -55,7 +55,7 @@ import { DomainEvent,
 import { randomUUID } from "crypto";
 
 // Tests can timeout in a CI pipeline so giving it leeway
-jest.setTimeout(20000)
+jest.setTimeout(10000)
 
 const logger: ILogger = new DefaultLogger('bc', 'appName', 'appVersion'); //TODO: parameterize the names here
 const messageTimeout = 2000;
@@ -829,504 +829,5 @@ describe("Tests for Outbound Command Event Handler", () => {
     // Check domain events published to kafka
     const hasAcceptPartyEvent = (domainEvents.find((e) => e.getName() === 'SDKOutboundBulkAutoAcceptPartyInfoRequestedDmEvt'));
     expect(hasAcceptPartyEvent).toBeTruthy();
-  });
-
-  test("9. Given inbound command event ProcessSDKOutboundBulkAcceptPartyInfo is received \
-        Then the logic should loop through individual transfer in the bulk request \
-          And update the individual transfer state to DISCOVERY_ACCEPTED or DISCOVERY_REJECTED based on the value in the incoming event \
-          And update the overall global state to DISCOVERY_ACCEPTANCE_COMPLETED \
-          And outbound event SDKOutboundBulkAcceptPartyInfoProcessed should be published", async () => {
-
-    //Publish this message so that it is stored internally in redis
-    const bulkTransactionId = randomUUID();
-    const bulkRequest: SDKSchemeAdapter.Outbound.V2_0_0.Types.bulkTransactionRequest = {
-        bulkHomeTransactionID: "string",
-        bulkTransactionId: bulkTransactionId,
-        options: {
-          onlyValidateParty: true,
-          autoAcceptParty: {
-            enabled: true
-          },
-          autoAcceptQuote: {
-            enabled: true,
-          },
-          skipPartyLookup: false,
-          synchronous: true,
-          bulkExpiration: "2016-05-24T08:38:08.699-04:00"
-        },
-        from: {
-          partyIdInfo: {
-            partyIdType: "MSISDN",
-            partyIdentifier: "16135551212",
-            fspId: "string",
-          },
-        },
-        individualTransfers: [
-          {
-            homeTransactionId: randomUUID(),
-            to: {
-              partyIdInfo: {
-                partyIdType: "MSISDN",
-                partyIdentifier: "16135551212"
-              },
-            },
-            amountType: "SEND",
-            currency: "USD",
-            amount: "123.45",
-          }
-        ]
-    }
-    const sampleCommandEventData: IProcessSDKOutboundBulkRequestCmdEvtData = {
-      bulkRequest,
-      timestamp: Date.now(),
-      headers: []
-    }
-    const processSDKOutboundBulkRequestMessageObj = new ProcessSDKOutboundBulkRequestCmdEvt(sampleCommandEventData);
-    await producer.sendCommandEvent(processSDKOutboundBulkRequestMessageObj);
-    await new Promise(resolve => setTimeout(resolve, messageTimeout));
-
-    // Command event for bulk party info request completed
-    const processSDKOutboundBulkPartyInfoRequestCompleteCommandEventData : IProcessSDKOutboundBulkPartyInfoRequestCompleteCmdEvtData = {
-      bulkId: bulkTransactionId,
-      timestamp: Date.now(),
-      headers: []
-    }
-    const processSDKOutboundBulkPartyInfoRequestCompleteCommandEventObj = new ProcessSDKOutboundBulkPartyInfoRequestCompleteCmdEvt(processSDKOutboundBulkPartyInfoRequestCompleteCommandEventData);
-    await producer.sendCommandEvent(processSDKOutboundBulkPartyInfoRequestCompleteCommandEventObj);
-    await new Promise(resolve => setTimeout(resolve, messageTimeout));
-
-    //Check that the global state of individual transfers in bulk to be RECEIVED
-    const bulkState = await bulkTransactionEntityRepo.load(bulkTransactionId);
-    expect(bulkState.state).toBe('DISCOVERY_COMPLETED');
-
-    // Check domain events published to kafka
-    const hasAcceptPartyEvent = (domainEvents.find((e) => e.getName() === 'SDKOutboundBulkAutoAcceptPartyInfoRequestedDmEvt'));
-    expect(hasAcceptPartyEvent).toBeTruthy();
-
-    // Get the randomly generated transferId for the callback
-    const previousIndividualTransfers = await bulkTransactionEntityRepo.getAllIndividualTransferIds(bulkTransactionId);
-
-    // Command event for bulk party info request completed
-    const processSDKOutboundBulkAcceptPartyInfoCommandEventData : IProcessSDKOutboundBulkAcceptPartyInfoCmdEvtData = {
-      bulkId: bulkTransactionId,
-      bulkTransactionContinuationAcceptParty: {
-        bulkHomeTransactionID: 'string',
-        individualTransfers: [
-          {
-            homeTransactionId: 'string',
-            transactionId: previousIndividualTransfers[0],
-            acceptParty: true
-          }
-        ]
-      },
-      timestamp: Date.now(),
-      headers: []
-    }
-    const processSDKOutboundBulkAcceptPartyInfoCommandEventObj = new ProcessSDKOutboundBulkAcceptPartyInfoCmdEvt(
-      processSDKOutboundBulkAcceptPartyInfoCommandEventData
-    );
-    await producer.sendCommandEvent(processSDKOutboundBulkAcceptPartyInfoCommandEventObj);
-    await new Promise(resolve => setTimeout(resolve, messageTimeout));
-
-    //Check that the global state of individual transfers in bulk to be RECEIVED
-    const bulkStateTwo = await bulkTransactionEntityRepo.load(bulkTransactionId);
-    expect(bulkStateTwo.state).toBe(BulkTransactionInternalState.DISCOVERY_ACCEPTANCE_COMPLETED);
-
-    // Get the randomly generated transferId for the callback
-    const afterIndividualTransfer = await bulkTransactionEntityRepo.getIndividualTransfer(bulkTransactionId,  previousIndividualTransfers[0]);
-    expect(afterIndividualTransfer.state).toBe(IndividualTransferInternalState.DISCOVERY_ACCEPTED);
-
-    // Check domain events published to kafka
-    console.log(domainEvents)
-    const hasSDKOutboundBulkAcceptPartyInfoProcessed = (domainEvents.find((e) => e.getName() === 'SDKOutboundBulkAcceptPartyInfoProcessedDmEvt'));
-    expect(hasSDKOutboundBulkAcceptPartyInfoProcessed).toBeTruthy();
-
-  });
-
-  // // TESTS FOR QUOTE PROCESSING
-
-  test("When Inbound command event ProcessSDKOutboundBulkQuotesRequest is received\
-        Then the logic should update the global state to AGREEMENT_PROCESSING, \
-        And create batches based on FSP that has DISCOVERY_ACCEPTED state \
-        And also has config maxEntryConfigPerBatch \
-        And publish BulkQuotesRequested per each batch \
-        And update the state of each batch to AGREEMENT_PROCESSING.", async () => {
-    //Publish this message so that it is stored internally in redis
-    const bulkTransactionId = randomUUID();
-    const bulkRequest: SDKSchemeAdapter.Outbound.V2_0_0.Types.bulkTransactionRequest = {
-        bulkHomeTransactionID: "string",
-        bulkTransactionId: bulkTransactionId,
-        options: {
-          onlyValidateParty: true,
-          autoAcceptParty: {
-            enabled: true
-          },
-          autoAcceptQuote: {
-            enabled: true,
-          },
-          skipPartyLookup: false,
-          synchronous: true,
-          bulkExpiration: "2016-05-24T08:38:08.699-04:00"
-        },
-        from: {
-          partyIdInfo: {
-            partyIdType: "MSISDN",
-            partyIdentifier: "16135551212",
-            fspId: "string",
-          },
-        },
-        individualTransfers: [
-          {
-            homeTransactionId: randomUUID(),
-            to: {
-              partyIdInfo: {
-                partyIdType: "MSISDN",
-                partyIdentifier: "16135551212"
-              },
-            },
-            amountType: "SEND",
-            currency: "USD",
-            amount: "123.45",
-          }
-        ]
-    }
-    const sampleCommandEventData: IProcessSDKOutboundBulkRequestCmdEvtData = {
-      bulkRequest,
-      timestamp: Date.now(),
-      headers: []
-    }
-    const processSDKOutboundBulkRequestMessageObj = new ProcessSDKOutboundBulkRequestCmdEvt(sampleCommandEventData);
-    await producer.sendCommandEvent(processSDKOutboundBulkRequestMessageObj);
-    await new Promise(resolve => setTimeout(resolve, messageTimeout));
-
-    const bulkPartyInfoRequestCommandEventData: IProcessSDKOutboundBulkPartyInfoRequestCmdEvtData = {
-      bulkId: bulkTransactionId,
-      timestamp: Date.now(),
-      headers: []
-    }
-    const bulkPartyInfoRequestCommandEventObj = new ProcessSDKOutboundBulkPartyInfoRequestCmdEvt(bulkPartyInfoRequestCommandEventData);
-    await producer.sendCommandEvent(bulkPartyInfoRequestCommandEventObj);
-
-    await new Promise(resolve => setTimeout(resolve, messageTimeout));
-    // Check the state in Redis
-    console.log('bulk id: ', bulkTransactionId);
-
-    const partyInfoRequestedDomainEvents = domainEvents.filter(domainEvent => domainEvent.getName() === 'PartyInfoRequestedDmEvt');
-
-    // Get the randomly generated transferId for the callback
-    const previousIndividualTransfersOne = await bulkTransactionEntityRepo.getAllIndividualTransferIds(bulkTransactionId);
-
-    const processPartyInfoCallbackMessageData: IProcessPartyInfoCallbackCmdEvtData = {
-      bulkId: partyInfoRequestedDomainEvents[0].getKey(),
-      content: {
-        transferId: previousIndividualTransfersOne[0],
-        partyResult: {
-          party: {
-              partyIdInfo: {
-                  partyIdType: 'MSISDN',
-                  partyIdentifier: '123456',
-                  fspId: 'receiverfsp'
-              }
-          },
-          currentState: 'COMPLETED'
-        },
-      },
-      timestamp: Date.now(),
-      headers: []
-    }
-    const processPartyInfoCallbackMessageObj = new ProcessPartyInfoCallbackCmdEvt(processPartyInfoCallbackMessageData);
-    await producer.sendCommandEvent(processPartyInfoCallbackMessageObj);
-    await new Promise(resolve => setTimeout(resolve, messageTimeout));
-
-    // Command event for bulk party info request completed
-    const processSDKOutboundBulkPartyInfoRequestCompleteCommandEventData : IProcessSDKOutboundBulkPartyInfoRequestCompleteCmdEvtData = {
-      bulkId: bulkTransactionId,
-      timestamp: Date.now(),
-      headers: []
-    }
-    const processSDKOutboundBulkPartyInfoRequestCompleteCommandEventObj = new ProcessSDKOutboundBulkPartyInfoRequestCompleteCmdEvt(processSDKOutboundBulkPartyInfoRequestCompleteCommandEventData);
-    await producer.sendCommandEvent(processSDKOutboundBulkPartyInfoRequestCompleteCommandEventObj);
-    await new Promise(resolve => setTimeout(resolve, messageTimeout));
-
-    //Check that the global state of individual transfers in bulk to be RECEIVED
-    const bulkState = await bulkTransactionEntityRepo.load(bulkTransactionId);
-    expect(bulkState.state).toBe('DISCOVERY_COMPLETED');
-
-    // Check domain events published to kafka
-    const hasAcceptPartyEvent = (domainEvents.find((e) => e.getName() === 'SDKOutboundBulkAutoAcceptPartyInfoRequestedDmEvt'));
-    expect(hasAcceptPartyEvent).toBeTruthy();
-
-    // Get the randomly generated transferId for the callback
-    const previousIndividualTransfersTwo = await bulkTransactionEntityRepo.getAllIndividualTransferIds(bulkTransactionId);
-
-    // Command event for bulk party info request completed
-    const processSDKOutboundBulkAcceptPartyInfoCommandEventData : IProcessSDKOutboundBulkAcceptPartyInfoCmdEvtData = {
-      bulkId: bulkTransactionId,
-      bulkTransactionContinuationAcceptParty: {
-        bulkHomeTransactionID: 'string',
-        individualTransfers: [
-          {
-            homeTransactionId: 'string',
-            transactionId: previousIndividualTransfersTwo[0],
-            acceptParty: true
-          }
-        ]
-      },
-      timestamp: Date.now(),
-      headers: []
-    }
-    const processSDKOutboundBulkAcceptPartyInfoCommandEventObj = new ProcessSDKOutboundBulkAcceptPartyInfoCmdEvt(
-      processSDKOutboundBulkAcceptPartyInfoCommandEventData
-    );
-    await producer.sendCommandEvent(processSDKOutboundBulkAcceptPartyInfoCommandEventObj);
-    await new Promise(resolve => setTimeout(resolve, messageTimeout));
-
-    //Check that the global state of individual transfers in bulk to be RECEIVED
-    const bulkStateTwo = await bulkTransactionEntityRepo.load(bulkTransactionId);
-    expect(bulkStateTwo.state).toBe(BulkTransactionInternalState.DISCOVERY_ACCEPTANCE_COMPLETED);
-
-    // Get the randomly generated transferId for the callback
-    const afterIndividualTransfer = await bulkTransactionEntityRepo.getIndividualTransfer(bulkTransactionId,  previousIndividualTransfersTwo[0]);
-    expect(afterIndividualTransfer.state).toBe(IndividualTransferInternalState.DISCOVERY_ACCEPTED);
-    console.log(afterIndividualTransfer);
-    // Check domain events published to kafka
-    const hasSDKOutboundBulkAcceptPartyInfoProcessed = (domainEvents.find((e) => e.getName() === 'SDKOutboundBulkAcceptPartyInfoProcessedDmEvt'));
-    expect(hasSDKOutboundBulkAcceptPartyInfoProcessed).toBeTruthy();
-
-    //xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
-    // Command event for bulk party info request completed
-    const processSDKOutboundBulkQuotesRequestCommandEventData : IProcessSDKOutboundBulkQuotesRequestCmdEvtData = {
-      bulkId: bulkTransactionId,
-      timestamp: Date.now(),
-      headers: []
-    }
-    const processSDKOutboundBulkQuotesRequestCommandEventObj = new ProcessSDKOutboundBulkQuotesRequestCmdEvt(
-      processSDKOutboundBulkQuotesRequestCommandEventData
-    );
-    await producer.sendCommandEvent(processSDKOutboundBulkQuotesRequestCommandEventObj);
-    await new Promise(resolve => setTimeout(resolve, messageTimeout));
-
-    //Check that the global state of individual transfers in bulk to be RECEIVED
-    const bulkStateThree = await bulkTransactionEntityRepo.load(bulkTransactionId);
-    expect(bulkStateThree.state).toBe(BulkTransactionInternalState.AGREEMENT_PROCESSING);
-
-    // Check domain events published to kafka
-    console.log(domainEvents);
-    const hasBulkQuotesRequested = (domainEvents.find((e) => e.getName() === 'BulkQuotesRequestedDmEvt'));
-    expect(hasBulkQuotesRequested).toBeTruthy();
-
-    const bulkBatchIds = await bulkTransactionEntityRepo.getAllBulkBatchIds(bulkTransactionId);
-    expect(bulkBatchIds[0]).toBeDefined();
-  });
-
-  test.only("Given Inbound command event ProcessBulkQuotesCallback for success requests \
-        Then the logic should update the individual batch state to AGREEMENT_PROCESSING, \
-        And create batches based on FSP that has DISCOVERY_ACCEPTED state \
-        And also has config maxEntryConfigPerBatch \
-        And publish BulkQuotesRequested per each batch \
-        And update the state of each batch to AGREEMENT_PROCESSING.", async () => {
-    //Publish this message so that it is stored internally in redis
-    const bulkTransactionId = randomUUID();
-    const bulkRequest: SDKSchemeAdapter.Outbound.V2_0_0.Types.bulkTransactionRequest = {
-        bulkHomeTransactionID: "string",
-        bulkTransactionId: bulkTransactionId,
-        options: {
-          onlyValidateParty: true,
-          autoAcceptParty: {
-            enabled: true
-          },
-          autoAcceptQuote: {
-            enabled: true,
-          },
-          skipPartyLookup: false,
-          synchronous: true,
-          bulkExpiration: "2016-05-24T08:38:08.699-04:00"
-        },
-        from: {
-          partyIdInfo: {
-            partyIdType: "MSISDN",
-            partyIdentifier: "16135551212",
-            fspId: "string",
-          },
-        },
-        individualTransfers: [
-          {
-            homeTransactionId: randomUUID(),
-            to: {
-              partyIdInfo: {
-                partyIdType: "MSISDN",
-                partyIdentifier: "16135551212"
-              },
-            },
-            amountType: "SEND",
-            currency: "USD",
-            amount: "123.45",
-          }
-        ]
-    }
-    const sampleCommandEventData: IProcessSDKOutboundBulkRequestCmdEvtData = {
-      bulkRequest,
-      timestamp: Date.now(),
-      headers: []
-    }
-    const processSDKOutboundBulkRequestMessageObj = new ProcessSDKOutboundBulkRequestCmdEvt(sampleCommandEventData);
-    await producer.sendCommandEvent(processSDKOutboundBulkRequestMessageObj);
-    await new Promise(resolve => setTimeout(resolve, messageTimeout));
-
-    const bulkPartyInfoRequestCommandEventData: IProcessSDKOutboundBulkPartyInfoRequestCmdEvtData = {
-      bulkId: bulkTransactionId,
-      timestamp: Date.now(),
-      headers: []
-    }
-    const bulkPartyInfoRequestCommandEventObj = new ProcessSDKOutboundBulkPartyInfoRequestCmdEvt(bulkPartyInfoRequestCommandEventData);
-    await producer.sendCommandEvent(bulkPartyInfoRequestCommandEventObj);
-
-    await new Promise(resolve => setTimeout(resolve, messageTimeout));
-    // Check the state in Redis
-    console.log('bulk id: ', bulkTransactionId);
-
-    const partyInfoRequestedDomainEvents = domainEvents.filter(domainEvent => domainEvent.getName() === 'PartyInfoRequestedDmEvt');
-
-    // Get the randomly generated transferId for the callback
-    const previousIndividualTransfersOne = await bulkTransactionEntityRepo.getAllIndividualTransferIds(bulkTransactionId);
-
-    const processPartyInfoCallbackMessageData: IProcessPartyInfoCallbackCmdEvtData = {
-      bulkId: partyInfoRequestedDomainEvents[0].getKey(),
-      content: {
-        transferId: previousIndividualTransfersOne[0],
-        partyResult: {
-          party: {
-              partyIdInfo: {
-                  partyIdType: 'MSISDN',
-                  partyIdentifier: '123456',
-                  fspId: 'receiverfsp'
-              }
-          },
-          currentState: 'COMPLETED'
-        },
-      },
-      timestamp: Date.now(),
-      headers: []
-    }
-    const processPartyInfoCallbackMessageObj = new ProcessPartyInfoCallbackCmdEvt(processPartyInfoCallbackMessageData);
-    await producer.sendCommandEvent(processPartyInfoCallbackMessageObj);
-    await new Promise(resolve => setTimeout(resolve, messageTimeout));
-
-    // Command event for bulk party info request completed
-    const processSDKOutboundBulkPartyInfoRequestCompleteCommandEventData : IProcessSDKOutboundBulkPartyInfoRequestCompleteCmdEvtData = {
-      bulkId: bulkTransactionId,
-      timestamp: Date.now(),
-      headers: []
-    }
-    const processSDKOutboundBulkPartyInfoRequestCompleteCommandEventObj = new ProcessSDKOutboundBulkPartyInfoRequestCompleteCmdEvt(processSDKOutboundBulkPartyInfoRequestCompleteCommandEventData);
-    await producer.sendCommandEvent(processSDKOutboundBulkPartyInfoRequestCompleteCommandEventObj);
-    await new Promise(resolve => setTimeout(resolve, messageTimeout));
-
-    //Check that the global state of individual transfers in bulk to be RECEIVED
-    const bulkState = await bulkTransactionEntityRepo.load(bulkTransactionId);
-    expect(bulkState.state).toBe('DISCOVERY_COMPLETED');
-
-    // Check domain events published to kafka
-    const hasAcceptPartyEvent = (domainEvents.find((e) => e.getName() === 'SDKOutboundBulkAutoAcceptPartyInfoRequestedDmEvt'));
-    expect(hasAcceptPartyEvent).toBeTruthy();
-
-    // Get the randomly generated transferId for the callback
-    const previousIndividualTransfersTwo = await bulkTransactionEntityRepo.getAllIndividualTransferIds(bulkTransactionId);
-
-    // Command event for bulk party info request completed
-    const processSDKOutboundBulkAcceptPartyInfoCommandEventData : IProcessSDKOutboundBulkAcceptPartyInfoCmdEvtData = {
-      bulkId: bulkTransactionId,
-      bulkTransactionContinuationAcceptParty: {
-        bulkHomeTransactionID: 'string',
-        individualTransfers: [
-          {
-            homeTransactionId: 'string',
-            transactionId: previousIndividualTransfersTwo[0],
-            acceptParty: true
-          }
-        ]
-      },
-      timestamp: Date.now(),
-      headers: []
-    }
-    const processSDKOutboundBulkAcceptPartyInfoCommandEventObj = new ProcessSDKOutboundBulkAcceptPartyInfoCmdEvt(
-      processSDKOutboundBulkAcceptPartyInfoCommandEventData
-    );
-    await producer.sendCommandEvent(processSDKOutboundBulkAcceptPartyInfoCommandEventObj);
-    await new Promise(resolve => setTimeout(resolve, messageTimeout));
-
-    //Check that the global state of individual transfers in bulk to be RECEIVED
-    const bulkStateTwo = await bulkTransactionEntityRepo.load(bulkTransactionId);
-    expect(bulkStateTwo.state).toBe(BulkTransactionInternalState.DISCOVERY_ACCEPTANCE_COMPLETED);
-
-    // Get the randomly generated transferId for the callback
-    const afterIndividualTransfer = await bulkTransactionEntityRepo.getIndividualTransfer(bulkTransactionId,  previousIndividualTransfersTwo[0]);
-    expect(afterIndividualTransfer.state).toBe(IndividualTransferInternalState.DISCOVERY_ACCEPTED);
-
-    // Check domain events published to kafka
-    const hasSDKOutboundBulkAcceptPartyInfoProcessed = (domainEvents.find((e) => e.getName() === 'SDKOutboundBulkAcceptPartyInfoProcessedDmEvt'));
-    expect(hasSDKOutboundBulkAcceptPartyInfoProcessed).toBeTruthy();
-
-    //xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
-    // Command event for bulk party info request completed
-    const processSDKOutboundBulkQuotesRequestCommandEventData : IProcessSDKOutboundBulkQuotesRequestCmdEvtData = {
-      bulkId: bulkTransactionId,
-      timestamp: Date.now(),
-      headers: []
-    }
-    const processSDKOutboundBulkQuotesRequestCommandEventObj = new ProcessSDKOutboundBulkQuotesRequestCmdEvt(
-      processSDKOutboundBulkQuotesRequestCommandEventData
-    );
-    await producer.sendCommandEvent(processSDKOutboundBulkQuotesRequestCommandEventObj);
-    await new Promise(resolve => setTimeout(resolve, messageTimeout));
-
-    //Check that the global state of individual transfers in bulk to be RECEIVED
-    const bulkStateThree = await bulkTransactionEntityRepo.load(bulkTransactionId);
-    expect(bulkStateThree.state).toBe(BulkTransactionInternalState.AGREEMENT_PROCESSING);
-
-    // Check domain events published to kafka
-    console.log(domainEvents);
-    const hasBulkQuotesRequested = (domainEvents.find((e) => e.getName() === 'BulkQuotesRequestedDmEvt'));
-    expect(hasBulkQuotesRequested).toBeTruthy();
-
-    const bulkBatchIds = await bulkTransactionEntityRepo.getAllBulkBatchIds(bulkTransactionId);
-    expect(bulkBatchIds[0]).toBeDefined();
-
-    //xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
-    // Command event for bulk party info request completed
-    // Get randomly generated quoteId
-    const preBulkBatch = await bulkTransactionEntityRepo.getBulkBatch(bulkTransactionId, bulkBatchIds[0]);
-
-    const bulkQuoteId = randomUUID();
-    const processBulkQuotesCallbackCommandEventData : IProcessBulkQuotesCallbackCmdEvtData = {
-      bulkId: bulkTransactionId,
-      content: {
-        batchId: bulkBatchIds[0],
-        bulkQuoteId: bulkQuoteId,
-        bulkQuotesResult: {
-          bulkQuoteId: bulkQuoteId,
-          currentState: 'COMPLETED',
-          individualQuoteResults: [
-            {
-              quoteId: preBulkBatch.bulkQuotesRequest.individualQuotes[0].quoteId,
-              transferAmount: {
-                currency: 'USD',
-                amount: '1',
-              },
-              ilpPacket: 'string',
-              condition: 'string'
-            }
-          ]
-        }
-      },
-      timestamp: Date.now(),
-      headers: []
-    }
-    const processBulkQuotesCallbackCommandEventObj = new ProcessBulkQuotesCallbackCmdEvt(
-      processBulkQuotesCallbackCommandEventData
-    );
-    await producer.sendCommandEvent(processBulkQuotesCallbackCommandEventObj);
-    await new Promise(resolve => setTimeout(resolve, messageTimeout));
-
   });
 });
